pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "agentseceval"
version = "0.1.0"
description = "Security benchmark evaluating isolation mechanisms for adversarial LLM agents"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2",
    "langchain-community>=0.2",
    "langchain-ollama>=0.2",
    "ollama>=0.2",
    "pyyaml>=6.0",
    "jsonschema>=4.0",
    "docker>=7.0",
    "pytest>=8.0",
    "rich>=13.0",
    "pandas>=2.0",
    "matplotlib>=3.8",
    "scipy>=1.12",
]

[project.scripts]
agentseceval = "agentseceval.experiments.orchestrator:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["agentseceval*"]
